Add docker configuration for each enviroment
