+++ .env_example
@@ -1,0 +1,15 @@
+# Database config
+POSTGRES_DB=backend
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
+POSTGRES_PORT=5432
+POSTGRES_HOST=db
+
+# Django config
+DJANGO_SECRET_KEY=4y7c-7+bja2@e1$%d@6v+#dir%70!dc7!7_04e2r(d%4$g7+id
+DJANGO_HASHID_FIELD_SALT='9q#3t$5gs9ob682b@(6^fdv2kg*0ztr(3doa((w&kyq!d8rbt^'
+DJANGO_SETTINGS_MODULE=settings.configs
+DJANGO_ALLOWED_HOSTS=localhost
+DJANGO_DEBUG=False
+DJANGO_CONFIGURATION=Local
+DJANGO_ADMIN_URL=admin
