+++ .gitignore
@@ -1,5 +1,6 @@
+db.sqlite3
+.env
+.idea/
 __pycache__/
 *.py[cod]
-.idea/
-db.sqlite3
 .pytest_cache/
