+++ docker-compose.yml
@@ -1,0 +1,22 @@
+version: "3"
+
+services:
+  backend:
+    build:
+      context: .
+      dockerfile: src/Dockerfile
+      args:
+        - ENV_NAME=local
+    env_file:
+      - .env
+    depends_on:
+      - db
+    ports:
+      - "8000:8000"
+    command: scripts/run_local.sh
+    volumes:
+      - ./src:/app/backend
+  db:
+    image: postgres:10.3-alpine
+    env_file:
+      - .env
