--- requirements.txt
@@ -1,4 +1,0 @@
-Django==2.0.2
-djangorestframework==3.7.7
-djangorestframework-jwt==1.11.0
-django-hashid-field==2.1.0
