+++ requirements/base.txt
@@ -1,0 +1,7 @@
+Django==2.0.2
+django-configurations==2.0
+django-hashid-field==2.1.0
+djangorestframework==3.7.7
+djangorestframework-jwt==1.11.0
+dj_database_url==0.5.0
+psycopg2==2.7.4
