+++ requirements/local.txt
@@ -1,0 +1,4 @@
+-r base.txt
+-r prod.txt
+-r qa.txt
+-r test.txt
