+++ requirements/qa.txt
@@ -1,0 +1,2 @@
+-r base.txt
+uwsgi==2.0.17
