+++ requirements/test.txt
@@ -1,0 +1,8 @@
+-r base.txt
+factory-boy==2.10.0
+flake8==3.5.0
+pytest==3.4.1
+pytest-django==3.1.2
+pytest-env==0.6.2
+pytest-factoryboy==2.0.1
+pytest-sugar==0.9.1
