--- requirements_test.txt
@@ -1,5 +1,0 @@
--r requirements.txt
-factory-boy==2.10.0
-pytest==3.4.1
-pytest-factoryboy==2.0.1
-pytest-django==3.1.2
