+++ src/manage.py
@@ -3,9 +3,10 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "restauth.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings.configs")
+    os.environ.setdefault('DJANGO_CONFIGURATION', 'Prod')
     try:
-        from django.core.management import execute_from_command_line
+        from configurations.management import execute_from_command_line
     except ImportError:
         # The above import may fail for some other reason. Ensure that the
         # issue is really that Django is missing to avoid masking other
