+++ src/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+python_files = tests.py test_*.py *_tests.py
+env =
+    DJANGO_SETTINGS_MODULE = settings.configs
+    DJANGO_CONFIGURATION = Test
