+++ src/restauth/apps.py
@@ -3,3 +3,4 @@ from django.apps import AppConfig
 
 class RestauthConfig(AppConfig):
     name = 'restauth'
+    verbose_name = 'Rest Auth'
