+++ src/scripts/run_local.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env sh
+
+set -x
+
+python manage.py migrate
+python manage.py runserver 0.0.0.0:8000 --settings=settings.configs --configuration=Local
