+++ src/scripts/run_prod.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env sh
+
+set -x
+
+python manage.py check --deploy
+python manage.py migrate
+uwsgi --ini $APP_DIR/uwsgi.ini
