+++ src/scripts/run_test.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+set -x
+
+pytest
