+++ src/settings/wsgi.py
@@ -9,8 +9,9 @@ https://docs.djangoproject.com/en/1.11/howto/deployment/wsgi/
 
 import os
 
-from django.core.wsgi import get_wsgi_application
+from configurations.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "restauth.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings.configs")
+os.environ.setdefault('DJANGO_CONFIGURATION', 'Prod')
 
 application = get_wsgi_application()
