+++ src/uwsgi.ini
@@ -1,0 +1,42 @@
+[uwsgi]
+
+# master
+master = true
+
+# processes
+enable-threads = true
+# With cheaper = 1 we activate the The uWSGI cheaper subsystem which allows to
+# dynamically scale the number of running workers (processes).
+cheaper = 1
+# The minimum load uWSGI will spawn just one workers.
+processes = %(%k + 1)
+threads = 2
+
+# sockets
+http-socket = :8000
+
+# wsgi
+chdir = /app/backend
+module = settings.wsgi
+
+# gracefully and sequentially reload workers when this file is touched
+touch-chain-reload = /tmp/backend.pid
+
+# customization
+die-on-term = true
+
+# clear environment on exit
+vacuum = true
+buffer-size = 32768
+
+# reload if rss memory is higher than specified megabytes
+reload-on-rss = 512
+
+# force the master to reload a worker if its rss memory is higher than specified megabytes
+evil-reload-on-rss = 1024
+
+# reload workers after the specified amount of managed requests
+max-requests = 100
+
+# reload workers after the specified amount of seconds
+max-worker-lifetime = 1800
